Fix searching on rec_name when the domain starts with an 'AND' or 'OR' operator.
Searching using the dot was implemented on the search() for both rec_name and
code. The problem is that the implementation does not consider the case when
the first argument contains an 'AND' or 'OR' operator.

When the user searches using the dot on a list view, the client prepends the
domain with 'AND' so the search did not work in that case.

The problem has been solved with:

- Implementing search_rec_name for the case of the 'rec_name' field.
- As the code field is not a function one, it has been implemented by
  overriding search_domain() and processing all the domain tree.